S05E04/requirements.txt
Flask==3.1.0
openai==1.82.1
python-dotenv==1.1.0
requests==2.31.0
Pillow==10.2.0
pydub==0.25.1